+++ src/App.js
@@ -1,4 +1,4 @@
-import './App.css';
+import './styles/App.css';
 
 import { Route, Routes, BrowserRouter } from 'react-router-dom';
 import Main from './pages/main';
